# Implement low-level API for jws, and incorporate it to JWT to achieve massive performance gains