--- a/internal/jwxio/BUILD.bazel
+++ b/internal/jwxio/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "jwxio",
+    srcs = ["jwxio.go"],
+    importpath = "github.com/lestrrat-go/jwx/v3/internal/jwxio",
+    visibility = ["//:__subpackages__"],
+)