--- a/jws/BUILD.bazel
+++ b/jws/BUILD.bazel
@@ -20,6 +20,8 @@ go_library(
         "options_gen.go",
         "rsa.go",
         "signer.go",
+        "sign_context.go",
+        "signature_builder.go",
         "verifier.go",
     ],
     importpath = "github.com/lestrrat-go/jwx/v3/jws",