Before this PR (from #1407)

```
goos: linux
goarch: amd64
pkg: github.com/lestrrat-go/jwx/bench/comparison
cpu: AMD Ryzen 9 7900X3D 12-Core Processor
BenchmarkGolangJWT_SignHS256-24                   674217              1819 ns/op            2298 B/op       35 allocs/op
BenchmarkGolangJWT_SignRS256-24                     1438            735844 ns/op            3332 B/op       32 allocs/op
BenchmarkGolangJWT_ParseHS256-24                  418237              2874 ns/op            2664 B/op       54 allocs/op
BenchmarkGolangJWT_ParseRS256-24                   48922             24458 ns/op            3888 B/op       58 allocs/op
BenchmarkGolangJWT_ParseWithClaims-24             393840              2837 ns/op            2328 B/op       52 allocs/op
BenchmarkGolangJWT_CreateAndParseHS256-24         237686              5029 ns/op            4965 B/op       89 allocs/op
BenchmarkJWX_JWTSignHS256-24                      157760              6826 ns/op            4673 B/op      117 allocs/op
BenchmarkJWX_JWTSignRS256-24                        1425            802451 ns/op           13661 B/op      170 allocs/op
BenchmarkJWX_JWTParseHS256-24                     107380             11070 ns/op           13397 B/op      200 allocs/op
BenchmarkJWX_JWTParseRS256-24                      32870             34509 ns/op           15352 B/op      214 allocs/op
BenchmarkJWX_JWTParseWithValidation-24            106356             10814 ns/op           13397 B/op      200 allocs/op
BenchmarkJWX_CreateAndParseHS256-24                67430             18058 ns/op           18369 B/op      325 allocs/op
BenchmarkJWX_JWSSignHS256-24                      253020              3994 ns/op            3254 B/op       71 allocs/op
BenchmarkJWX_JWSParseHS256-24                     301191              3702 ns/op            4048 B/op       73 allocs/op
PASS
ok      github.com/lestrrat-go/jwx/bench/comparison     19.600s
```

With this PR, we can see 10%~  performance gain based on allocation. In some common cases we can now achieve 70%~.

```
goos: linux
goarch: amd64
pkg: github.com/lestrrat-go/jwx/bench/comparison
cpu: AMD Ryzen 9 7900X3D 12-Core Processor          
BenchmarkGolangJWT_SignHS256-24                   607450              1880 ns/op            2298 B/op         35 allocs/op
BenchmarkGolangJWT_SignRS256-24                     1663            736448 ns/op            3331 B/op         32 allocs/op
BenchmarkGolangJWT_ParseHS256-24                  389757              3028 ns/op            2664 B/op         54 allocs/op
BenchmarkGolangJWT_ParseRS256-24                   47256             24775 ns/op            3888 B/op         58 allocs/op
BenchmarkGolangJWT_ParseWithClaims-24             380250              2978 ns/op            2328 B/op         52 allocs/op
BenchmarkGolangJWT_CreateAndParseHS256-24         228819              5133 ns/op            4965 B/op         89 allocs/op
BenchmarkJWX_JWTSignHS256-24                      392856              2934 ns/op            1803 B/op         37 allocs/op
BenchmarkJWX_JWTSignRS256-24                        1576            797445 ns/op            9340 B/op         88 allocs/op
BenchmarkJWX_JWTParseHS256-24                     144380              8473 ns/op            9853 B/op        126 allocs/op
BenchmarkJWX_JWTParseRS256-24                      37928             32123 ns/op           11845 B/op        142 allocs/op
BenchmarkJWX_JWTParseWithValidation-24            128775              8157 ns/op            9853 B/op        126 allocs/op
BenchmarkJWX_CreateAndParseHS256-24               104121             12017 ns/op           11960 B/op        171 allocs/op
BenchmarkJWX_JWSSignHS256-24                      320575              3583 ns/op            2705 B/op         56 allocs/op
BenchmarkJWX_JWSParseHS256-24                     325221              3566 ns/op            3715 B/op         61 allocs/op
PASS
ok      github.com/lestrrat-go/jwx/bench/comparison     18.061s
```